feat: Improve rendering controls UX and fix DOF issues
- Change rendering control panel toggle key from 'A' to 'R' for better ergonomics
- Fix DOF focus distance by converting to normalized depth buffer space (0-1)
- Add comprehensive tooltips to all rendering control parameters
- Update colors to float32 for full HDR support throughout the pipeline
- Update documentation to reflect all changes
- Fix test expecting uint8 colors (now float32 for HDR)

+++ packages/luxar/src/luxar/tests/test_scene_structure.py
@@ -29,4 +29,4 @@ def test_random_demo_roundtrip(tmp_path):
 
     col = grp["colors"]
     assert col.shape == (n, 3)
-    assert col.dtype == np.uint8
+    assert col.dtype == np.float32  # HDR colors are now float32
